backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic[email]==2.5.3
python-multipart==0.0.6
email-validator==2.1.0
